chore: Update ScalariformDemoPage to include userBus and renderUser function
The code changes in ScalariformDemoPage include the addition of the userBus variable and the renderUser function. The userBus is an EventBus that emits User events, and the renderUser function renders the User data on the page. This update enhances the functionality of the page by allowing the display of user information and handling user events.